Widen numeric identifiers to long

+++ src/main/java/com/github/zafarkhaja/semver/Version.java
@@ -273,7 +273,7 @@ public class Version implements Comparable<Version> {
      * @throws IllegalArgumentException if the argument is a negative number
      * @since 0.10.0
      */
-    public static Version of(int major) {
+    public static Version of(long major) {
         return Version.of(major, 0, 0, null, null);
     }
 
@@ -288,7 +288,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, String preRelease) {
+    public static Version of(long major, String preRelease) {
         return Version.of(major, 0, 0, preRelease, null);
     }
 
@@ -305,7 +305,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, String preRelease, String build) {
+    public static Version of(long major, String preRelease, String build) {
         return Version.of(major, 0, 0, preRelease, build);
     }
 
@@ -318,7 +318,7 @@ public class Version implements Comparable<Version> {
      * @throws IllegalArgumentException if any of the arguments is a negative number
      * @since 0.10.0
      */
-    public static Version of(int major, int minor) {
+    public static Version of(long major, long minor) {
         return Version.of(major, minor, 0, null, null);
     }
 
@@ -334,7 +334,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, int minor, String preRelease) {
+    public static Version of(long major, long minor, String preRelease) {
         return Version.of(major, minor, 0, preRelease, null);
     }
 
@@ -352,7 +352,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, int minor, String preRelease, String build) {
+    public static Version of(long major, long minor, String preRelease, String build) {
         return Version.of(major, minor, 0, preRelease, build);
     }
 
@@ -366,7 +366,7 @@ public class Version implements Comparable<Version> {
      * @throws IllegalArgumentException if any of the arguments is a negative number
      * @since 0.10.0
      */
-    public static Version of(int major, int minor, int patch) {
+    public static Version of(long major, long minor, long patch) {
         return Version.of(major, minor, patch, null, null);
     }
 
@@ -383,7 +383,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, int minor, int patch, String preRelease) {
+    public static Version of(long major, long minor, long patch, String preRelease) {
         return Version.of(major, minor, patch, preRelease, null);
     }
 
@@ -402,7 +402,7 @@ public class Version implements Comparable<Version> {
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
      * @since 0.10.0
      */
-    public static Version of(int major, int minor, int patch, String preRelease, String build) {
+    public static Version of(long major, long minor, long patch, String preRelease, String build) {
         return new Version(
             new NormalVersion(major, minor, patch),
             preRelease == null ? MetadataVersion.NULL : VersionParser.parsePreRelease(preRelease),
@@ -419,7 +419,7 @@ public class Version implements Comparable<Version> {
     }
 
     /**
-     * @deprecated forRemoval since 0.10.0, use {@link Version#of(int)} instead.
+     * @deprecated forRemoval since 0.10.0, use {@link Version#of(long)} instead.
      */
     @Deprecated
     public static Version forIntegers(int major) {
@@ -427,7 +427,7 @@ public class Version implements Comparable<Version> {
     }
 
     /**
-     * @deprecated forRemoval since 0.10.0, use {@link Version#of(int, int)} instead.
+     * @deprecated forRemoval since 0.10.0, use {@link Version#of(long, long)} instead.
      */
     @Deprecated
     public static Version forIntegers(int major, int minor) {
@@ -435,7 +435,7 @@ public class Version implements Comparable<Version> {
     }
 
     /**
-     * @deprecated forRemoval since 0.10.0, use {@link Version#of(int, int, int)} instead.
+     * @deprecated forRemoval since 0.10.0, use {@link Version#of(long, long, long)} instead.
      */
     @Deprecated
     public static Version forIntegers(int major, int minor, int patch) {
@@ -478,6 +478,7 @@ public class Version implements Comparable<Version> {
      * Increments the major version.
      *
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the major version number overflows
      */
     public Version incrementMajorVersion() {
         return new Version(normal.incrementMajor());
@@ -488,6 +489,7 @@ public class Version implements Comparable<Version> {
      *
      * @param preRelease the pre-release version to append
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the major version number overflows
      * @throws IllegalArgumentException if the input string is {@code NULL} or empty
      * @throws ParseException when invalid version string is provided
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
@@ -503,6 +505,7 @@ public class Version implements Comparable<Version> {
      * Increments the minor version.
      *
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the minor version number overflows
      */
     public Version incrementMinorVersion() {
         return new Version(normal.incrementMinor());
@@ -513,6 +516,7 @@ public class Version implements Comparable<Version> {
      *
      * @param preRelease the pre-release version to append
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the minor version number overflows
      * @throws IllegalArgumentException if the input string is {@code NULL} or empty
      * @throws ParseException when invalid version string is provided
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
@@ -528,6 +532,7 @@ public class Version implements Comparable<Version> {
      * Increments the patch version.
      *
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the patch version number overflows
      */
     public Version incrementPatchVersion() {
         return new Version(normal.incrementPatch());
@@ -538,6 +543,7 @@ public class Version implements Comparable<Version> {
      *
      * @param preRelease the pre-release version to append
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the patch version number overflows
      * @throws IllegalArgumentException if the input string is {@code NULL} or empty
      * @throws ParseException when invalid version string is provided
      * @throws UnexpectedCharacterException is a special case of {@code ParseException}
@@ -553,6 +559,7 @@ public class Version implements Comparable<Version> {
      * Increments the pre-release version.
      *
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the numeric identifier overflows
      */
     public Version incrementPreReleaseVersion() {
         return new Version(normal, preRelease.increment());
@@ -562,6 +569,7 @@ public class Version implements Comparable<Version> {
      * Increments the build metadata.
      *
      * @return a new instance of the {@code Version} class
+     * @throws ArithmeticException if the numeric identifier overflows
      */
     public Version incrementBuildMetadata() {
         return new Version(normal, preRelease, build.increment());
@@ -598,7 +606,7 @@ public class Version implements Comparable<Version> {
      *
      * @return the major version number
      */
-    public int getMajorVersion() {
+    public long getMajorVersion() {
         return normal.getMajor();
     }
 
@@ -607,7 +615,7 @@ public class Version implements Comparable<Version> {
      *
      * @return the minor version number
      */
-    public int getMinorVersion() {
+    public long getMinorVersion() {
         return normal.getMinor();
     }
 
@@ -616,7 +624,7 @@ public class Version implements Comparable<Version> {
      *
      * @return the patch version number
      */
-    public int getPatchVersion() {
+    public long getPatchVersion() {
         return normal.getPatch();
     }
